pyproject.toml
[project]
name = "jaacountable-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosql>=13.4",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "google-adk[eval]==1.19.0",
    "greenlet>=3.2.3",
    "litellm>=1.80.11",
    "lxml>=6.0.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.3.0",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "testcontainers[postgres]>=4.13.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "contract: marks tests that validate external service contracts (make live HTTP requests)",
    "external: marks tests that make external calls (integration + contract tests)",
]

